# Update header parser in Secure Cell master key API