--- a/src/themis/secure_cell_alg.h
+++ b/src/themis/secure_cell_alg.h
@@ -88,4 +88,21 @@
 #define THEMIS_SYM_IV_LENGTH 16
 #endif
 
+static inline bool soter_alg_reserved_bits_valid(uint32_t alg)
+{
+    static const uint32_t used_bits = SOTER_SYM_KEY_LENGTH_MASK | SOTER_SYM_PADDING_MASK
+                                      | SOTER_SYM_KDF_MASK | SOTER_SYM_ALG_MASK;
+    return (alg & ~used_bits) == 0;
+}
+
+static inline size_t soter_alg_key_length(uint32_t alg)
+{
+    return (alg & SOTER_SYM_KEY_LENGTH_MASK) / 8;
+}
+
+static inline uint32_t soter_alg_kdf(uint32_t alg)
+{
+    return alg & SOTER_SYM_KDF_MASK;
+}
+
 #endif /* THEMIS_SECURE_CELL_ALG_H */