Passphrase API added in #577 has added quite a few new utilities and more robust and readable approach to parsing Secure Cell headers. Fuzzing with AFL in #579 and #583 has uncovered a bunch of sleeper issues in parsing code. This PR incorporates those improvements and updates master key API implementation to resolve all those issues.

First bunch of commits unifies master key implementation with passphrase one, now they are more or less similar, with differences only in KDF processing and header handling. Maybe later we will be able to reduce code duplication even further.

Some next commits resolve miscellaneous issues found in #583 for passphrase API which were also actual for master key API. This should unbreak fuzzing tests performed by #579.

And there are some minor commits with “collateral damage”: extracted common code from passphrase API implementation and some fixups in tests to expect different error codes.

## Checklist

- [X] Change is covered by automated tests
- [x] Benchmark results are attached (manually compare with master)
- [X] The [coding guidelines] are followed
- [X] ~~Public API has proper documentation~~ (no changes)
- [X] ~~Changelog is updated~~ (will update in bulk before release)

[coding guidelines]: https://github.com/cossacklabs/themis/blob/master/CONTRIBUTING.md