@ilammy 2020-02-12
Since I did not teach our CI to compare benchmark results yet, I'm filling it for this job.

Here are reports for [master](https://app.circleci.com/jobs/github/cossacklabs/themis/8598/artifacts) and [this PR](https://app.circleci.com/jobs/github/cossacklabs/themis/8623/artifacts), they check only processing of 4 KB of data.

| | Encryption | Decryption |
| -- | -- | -- |
| Before | 4.38 µs ± 0.73 µs | 3.10 µs ± 0.26 µs |
| After | 4.20 µs ± 0.18 µs | 3.01 µs ± 0.28 µs |

So it’s even a slight improvement but within the variance so I doubt it is statistically significant.

Performance optimization is not a goal of this PR so I won‘t be doing measurements manually. I’m fine with not having any obvious degradation.